+++ cogs/util.py
@@ -5,6 +5,9 @@ class UtilCog(commands.Cog):
     def __init__(self, client):
         self.client = client
 
+    @commands.command()
+    async def echo(self, ctx, *, arg):
+        await ctx.send(f"{arg}")
 
 def setup(client):
     client.add_cog(UtilCog(client))