+++ changelog.md
@@ -1,6 +1,22 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2020-08-08
+### Added
+- Bot reacts with 👋 to messages that start with words like "hello" and "sup".
+- Bot responds to "stfu" with "no u".
+- Easter cog, currently does nothing and isn't loaded by the main script.
+- Info cog.
+- Github command in info cog
+- License command with pretty embed in info cog
+- Echo command in util cog
+### Changed
+- Bot status 🤭
+- Updated README
+- Updated changelog
+### Removed
+- Base cog, along with 0 features
+
 ## [0.0.1] - 2020-08-07 - Initial Release
 ### Features
 - A manager that starts the bot, allowing for starting and stopping the bot within discord.